Decompose queries from conversation list
